requirements.txt
flask
tensorflow
PyJWT
python-dotenv